promise: fix signed/unsigned warning

+++ promise.cc
@@ -15,7 +15,7 @@ bool Promise::increment_threads(thread_id_t tid) {
 	unsigned int sync_size=synced_thread.size();
 	int promise_tid=id_to_int(read->get_tid());
 	for(unsigned int i=1;i<model->get_num_threads();i++) {
-		if ((i >= sync_size || !synced_thread[i]) && ( i != promise_tid ) && (enabled[i] != THREAD_DISABLED)) {
+		if ((i >= sync_size || !synced_thread[i]) && ( (int)i != promise_tid ) && (enabled[i] != THREAD_DISABLED)) {
 			return false;
 		}
 	}
